+++ src/scalib_ext/scalib/src/lib.rs
@@ -1,4 +1,6 @@
 pub mod belief_propagation;
+pub(crate) mod bp_compute;
+pub mod factor_graph;
 pub mod lda;
 pub(crate) mod matrixmul;
 pub mod mttest;